♻️ refactor(customer): improve error handling and finalize service/controller test logic

+++ services/customer/src/main/java/com/azouz/ecommerce/customer/service/CustomerServiceImpl.java
@@ -55,6 +55,9 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public void deleteCustomer(String id) {
+        if (!repository.existsById(id)) {
+            throw new CustomerNotFoundException();
+        }
         repository.deleteById(id);
     }
